requirements.txt
augly>=0.1.10
aiohttp
bottle
cheroot
PyJWT>=2.0.0
numpy
pymysql
sagemaker
six
SQLAlchemy==1.3.23
torch
transformers
werkzeug
boto3==1.24.56
yoyo-migrations
pandas==1.3.0
sacrebleu>=2.0.0
sklearn
git+https://github.com/facebookresearch/dynalab.git
sentencepiece
spacy
pyyaml
textflint
ujson>=4.2.0
python-dotenv
